rename pages to jsx with indexes
