+++ client/pages/contact/Contact.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Contact = () => {
+    return (
+        <div>
+            Contact
+        </div>
+    )
+}
+
+export default Contact;
