+++ client/pages/contact/index.js
@@ -1,11 +1,3 @@
-import React from 'react'
-
-const Contact = () => {
-    return (
-        <div>
-            Contact
-        </div>
-    )
-}
+import Contact from "./Contact";
 
 export default Contact;
