+++ client/pages/leagues/index.js
@@ -1,23 +1,3 @@
-import React from "react";
-import Button from "@material-ui/core/Button";
-import { LeagueStyled } from "../../styles";
-
-const Leagues = () => {
-  return (
-    <LeagueStyled>
-      <h1 className="title">Leagues</h1>
-      <div className="header">
-        <Button
-          size="large"
-          variant="contained"
-          color="primary"
-          className="create-button"
-        >
-          Create Team
-        </Button>
-      </div>
-    </LeagueStyled>
-  );
-};
+import Leagues from "./Leagues";
 
 export default Leagues;
